+++ app/src/androidTest/java/com/github/techisfun/onelinecalendar/CustomAndroidTestRunner.java
@@ -1,8 +1,8 @@
 package com.github.techisfun.onelinecalendar;
 
 import android.os.Bundle;
-import android.support.test.runner.AndroidJUnitRunner;
 
+import androidx.test.runner.AndroidJUnitRunner;
 import com.facebook.testing.screenshot.ScreenshotRunner;
 
 /**